Reorganize the source.
