+++ src/ui/scaffolding/app.js
