+++ src/ui/task-components/editor/progress-bar.js
